add third group parser, create library
