+++ kn15/__init__.py
@@ -1,5 +1,6 @@
-from .kn15 import KN15, decode, KN15Error
+from .kn15 import KN15, decode
 from .hydra.daily_standard import StandardObservation
+from .hydra.stage_and_flow import StageAndFlow
 
 
 __version__ = '1.1.0'
